rename `~/.-shrc` to `~/.shrc`
the reason for the hyphen is that i was originally concerned
about random shells executing the file when i didn't want them to.
back then, i was only targetting zsh and bash, and therefore
`~/.-shrc` was not yet suitable for any other shells.
nowadays, `~/.-shrc` is has better shell compatibility,
and i'm not actually aware of any shells that would execute
`~/.shrc` by default -- `$ENV` is typically required to be set.